+++ packages/platform-android/src/commands/runAndroid/__tests__/getLaunchPackageName.test.js
@@ -4,46 +4,11 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  *
+ * @flow
  */
 
-import fs from 'fs';
 import getLaunchPackageName from '../getLaunchPackageName';
-
-jest.mock('fs');
-jest.mock('path');
-
-function createMocks(useFlavor) {
-  const actualFs = jest.requireActual('fs');
-  const actualPath = jest.requireActual('path');
-
-  fs.readFileSync = jest.fn(filename => {
-    switch (filename) {
-      case actualPath.join('app', 'build.gradle'):
-        return actualFs.readFileSync(
-          actualPath.join(
-            __dirname,
-            '..',
-            '__fixtures__',
-            useFlavor ? 'sampleBuildWithFlavor.gradle' : 'sampleBuild.gradle',
-          ),
-          'utf8',
-        );
-      // Use default case to catch generated debug manifest
-      default:
-        return actualFs.readFileSync(
-          actualPath.join(
-            __dirname,
-            '..',
-            '__fixtures__',
-            useFlavor
-              ? 'sampleGeneratedDemoDebugManifest.xml'
-              : 'sampleGeneratedDebugManifest.xml',
-          ),
-          'utf8',
-        );
-    }
-  });
-}
+import {createBuildGradleMocks} from './testHelpers';
 
 describe('run-android::getLaunchPackageName', () => {
   beforeEach(() => {
@@ -51,7 +16,7 @@ describe('run-android::getLaunchPackageName', () => {
   });
 
   it('returns a well formed package name for unflavored build type', () => {
-    createMocks(/* useFlavor: */ false);
+    createBuildGradleMocks(false);
 
     const basePackageName = 'com.mycompany.app';
 
@@ -59,7 +24,7 @@ describe('run-android::getLaunchPackageName', () => {
   });
 
   it('returns a well formed package for flavored build type', () => {
-    createMocks(/* useFlavor: */ true);
+    createBuildGradleMocks(true);
 
     const basePackageName = 'com.mycompany.app';
